+++ common/config/networks.handlebars.json
@@ -4,7 +4,8 @@
     "chainName": "mainnet",
     "rpcUrls": [
       "https://eth-mainnet.g.alchemy.com/v2/{{env.ALCHEMY_API_KEY}}",
-      "https://mainnet.infura.io/v3/{{env.INFURA_API_KEY}}"
+      "https://mainnet.infura.io/v3/{{env.INFURA_API_KEY}}",
+      "https://rpc.ankr.com/eth"
     ],
     "nativeCurrency": {
       "name": "Ether",
@@ -40,7 +41,8 @@
     "chainName": "arbitrum",
     "rpcUrls": [
       "https://arb-mainnet.g.alchemy.com/v2/{{env.ALCHEMY_API_KEY_ARBITRUM}}",
-      "https://arbitrum-mainnet.infura.io/v3/{{env.INFURA_API_KEY}}"
+      "https://arbitrum-mainnet.infura.io/v3/{{env.INFURA_API_KEY}}",
+      "https://rpc.ankr.com/arbitrum"
     ],
     "nativeCurrency": {
       "name": "Ether",
@@ -68,6 +70,7 @@
     "rpcUrls": [
       "https://opt-mainnet.g.alchemy.com/v2/{{env.ALCHEMY_API_KEY_OPTIMISM}}",
       "https://optimism-mainnet.infura.io/v3/{{env.INFURA_API_KEY}}",
+      "https://rpc.ankr.com/optimism",
       "https://mainnet.optimism.io"
     ],
     "nativeCurrency": {
@@ -95,6 +98,7 @@
     "rpcUrls": [
       "https://polygon-mainnet.g.alchemy.com/v2/{{env.ALCHEMY_API_KEY_POLYGON}}",
       "https://polygon-mainnet.infura.io/v3/{{env.INFURA_API_KEY}}",
+      "https://rpc.ankr.com/polygon",
       "https://polygon-rpc.com/"
     ],
     "nativeCurrency": {
@@ -133,6 +137,7 @@
     "chainName": "avalanche",
     "rpcUrls": [
       "https://avalanche-mainnet.infura.io/v3/{{env.INFURA_API_KEY}}",
+      "https://rpc.ankr.com/avalanche",
       "https://api.avax.network/ext/bc/C/rpc"
     ],
     "nativeCurrency": {
@@ -161,6 +166,7 @@
     "chainName": "xdai",
     "rpcUrls": [
       "https://rpc.ankr.com/gnosis",
+      "https://rpc.gnosischain.com",
       "https://rpc.xdaichain.com",
       "https://xdai.poanetwork.dev/",
       "https://dai.poa.network/"
@@ -190,6 +196,7 @@
     "chainName": "fantom",
     "rpcUrls": [
       "https://rpc.ftm.tools/",
+      "https: //rpc.ankr.com/fantom",
       "https://rpcapi.fantom.network"
     ],
     "nativeCurrency": {
@@ -217,6 +224,7 @@
     "chainName": "bsc",
     "rpcUrls": [
       "https://bsc-dataseed1.binance.org",
+      "https://rpc.ankr.com/bsc",
       "https://bsc-dataseed.binance.org",
       "https://bsc-dataseed1.defibit.io",
       "https://bsc-dataseed1.ninicoin.io"
@@ -246,6 +254,7 @@
     "chainName": "bsctestnet",
     "rpcUrls": [
       "https://data-seed-prebsc-1-s1.binance.org:8545",
+      "https://rpc.ankr.com/bsc_testnet_chapel",
       "https://data-seed-prebsc-2-s1.binance.org:8545",
       "https://data-seed-prebsc-1-s2.binance.org:8545"
     ],
@@ -296,6 +305,7 @@
     "chainId": 5,
     "chainName": "goerli",
     "rpcUrls": [
+      "https://rpc.ankr.com/eth_goerli",
       "https://goerli.infura.io/v3/{{env.INFURA_API_KEY}}"
     ],
     "nativeCurrency": {
@@ -326,6 +336,7 @@
     "chainName": "mumbai",
     "rpcUrls": [
       "https://polygon-mumbai.infura.io/v3/{{env.INFURA_API_KEY}}",
+      "https://rpc.ankr.com/polygon_mumbai",
       "https://matic-mumbai.chainstacklabs.com",
       "https://rpc-mumbai.maticvigil.com/v1"
     ],
@@ -363,7 +374,8 @@
     "chainId": 4002,
     "chainName": "fantomtestnet",
     "rpcUrls": [
-      "https://rpc.testnet.fantom.network"
+      "https://rpc.testnet.fantom.network",
+      "https://rpc.ankr.com/fantom_testnet"
     ],
     "nativeCurrency": {
       "name": "FTM",
@@ -387,6 +399,7 @@
     "chainName": "fuji",
     "rpcUrls": [
       "https://avalanche-fuji.infura.io/v3/{{env.INFURA_API_KEY}}",
+      "https://rpc.ankr.com/avalanche_fuji",
       "https://api.avax-test.network/ext/bc/C/rpc"
     ],
     "nativeCurrency": {
