+++ svelte/components/Global/Content.svelte
@@ -2,7 +2,7 @@
   import { setContext } from "svelte";
   import { writable } from "svelte/store";
 
-  import Nft from "../Nft/Nft.svelte";
+  import Nft from "../Nft/NftOld.svelte";
   import NftsList from "../NftsList/NftsList.svelte";
 
   ////////////////////////////////////////////////////////////////////
